requirements.txt
appdirs==1.4.3
asgiref==3.3.1
boto3==1.17.32
botocore==1.20.32
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.3.1
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jmespath==0.10.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.4
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.6
six==1.14.0
sqlparse==0.4.1
urllib3==1.25.8
webencodings==0.5.1